Add `vcpkg` build support (#366)
Closes #547.

Adds `vcpkg` build support to build a fully statically linked extension
for distribution:
```
ldd build/vcpkg-release/extension/sirius/sirius.duckdb_extension
          linux-vdso.so.1 (0x00007fb1b8676000)
          libdl.so.2 => /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib/libdl.so.2 (0x00007fb1a37f9000)
          librt.so.1 => /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib/librt.so.1 (0x00007fb1a37f4000)
          libpthread.so.0 => /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib/libpthread.so.0 (0x00007fb1a37ef000)
          libstdc++.so.6 => not found
          libm.so.6 => /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib/libm.so.6 (0x00007fb1a36f7000)
          libgcc_s.so.1 => /home/matthijs/sirius-vcpkg/.pixi/envs/vcpkg/lib/libgcc_s.so.1 (0x00007fb1a36c8000)
          libc.so.6 => /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib/libc.so.6 (0x00007fb1a3000000)
          /nix/store/l0l2ll1lmylczj1ihqn351af2kyp5x19-glibc-2.42-51/lib64/ld-linux-x86-64.so.2 (0x00007fb1b8678000)
```

---------

+++ src/op/sirius_physical_hash_join.cpp
@@ -501,7 +501,7 @@ std::unique_ptr<operator_data> sirius_physical_hash_join::get_next_task_input_da
     SIRIUS_LOG_WARN(fmt::format(
       "In sirius_physical_hash_join:get_next_task_input_data_for_build_probe: invalid hash table "
       "build state {} in operator {}",
-      _hash_table_build_state,
+      static_cast<int>(_hash_table_build_state),
       this->get_operator_id()));
     return nullptr;
   }
